requirements.txt
streamlit
tensorflow
keras
keras-tuner
numpy
pandas
matplotlib
Pillow
pypdf
openai